+++ crates/nodes/src/cs/light/info.rs
@@ -28,7 +28,7 @@ const LIGHT_BBOX: BoundingBox = BoundingBox {
         z: -1.0,
     },
 };
-pub fn assert_variants(node: NodeVariantsCs, offset: u32) -> Result<NodeVariantCs> {
+pub fn assert_variants(node: NodeVariantsCs, offset: usize) -> Result<NodeVariantCs> {
     assert_that!("light name", node.name eq LIGHT_NAME, offset + 0)?;
     assert_that!("light flags", node.flags == LIGHT_FLAGS, offset + 36)?;
     assert_that!("light field 040", node.unk040 == 0x00000000, offset + 40)?;
