+++ crates/nodes/src/cs/world/info.rs
@@ -15,7 +15,7 @@ const WORLD_FLAGS: NodeBitFlagsCs = NodeBitFlagsCs::from_bits_truncate(
         | NodeBitFlagsCs::UNK24.bits(),
 );
 
-pub fn assert_variants(node: NodeVariantsCs, offset: u32) -> Result<NodeVariantCs> {
+pub fn assert_variants(node: NodeVariantsCs, offset: usize) -> Result<NodeVariantCs> {
     assert_that!("world name", node.name eq WORLD_NAME, offset + 0)?;
     assert_that!("world flags", node.flags == WORLD_FLAGS, offset + 36)?;
     assert_that!("world field 040", node.unk040 == 0x00000000, offset + 40)?;
