+++ crates/nodes/src/mw/empty/info.rs
@@ -25,7 +25,7 @@ const VARIABLE_FLAGS: NodeBitFlags = NodeBitFlags::from_bits_truncate(
     | 0,
 );
 
-pub fn assert_variants(node: NodeVariantsMw, offset: u32) -> Result<NodeVariantMw> {
+pub fn assert_variants(node: NodeVariantsMw, offset: usize) -> Result<NodeVariantMw> {
     // cannot assert name
     let const_flags = node.flags & !VARIABLE_FLAGS;
     assert_that!("empty flags", const_flags == ALWAYS_PRESENT, offset + 36)?;
