+++ crates/nodes/src/rc/display/info.rs
@@ -7,7 +7,7 @@ use mech3ax_common::{assert_that, Result};
 
 const DISPLAY_NAME: &str = "display";
 
-pub fn assert_variants(node: NodeVariantsRc, offset: u32) -> Result<NodeVariantRc> {
+pub fn assert_variants(node: NodeVariantsRc, offset: usize) -> Result<NodeVariantRc> {
     assert_that!("display name", node.name eq DISPLAY_NAME, offset + 0)?;
     assert_that!(
         "display flags",
