+++ crates/nodes/src/rc/lod/data.rs
@@ -26,7 +26,7 @@ struct LodRcC {
 }
 impl_as_bytes!(LodRcC, 80);
 
-fn assert_lod(lod: &LodRcC, offset: u32) -> Result<(bool, Range, f32, Option<u32>)> {
+fn assert_lod(lod: &LodRcC, offset: usize) -> Result<(bool, Range, f32, Option<u32>)> {
     let level = assert_that!("lod level", bool lod.level, offset + 0)?;
     assert_that!("lod range near sq", 0.0 <= lod.range_near_sq <= 1000.0 * 1000.0, offset + 4)?;
     let range_near = lod.range_near_sq.sqrt();
